End of seciton 10 - validation for api

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
 using System;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,19 +7,24 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    //Unit就是返回 void 的意思
+    public class Command : IRequest<Result<Unit>>
     {
         public required string Id { get; set; }
     }
-    public class Handler(AppDbContext context) : IRequestHandler<Command>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var activity = await context.Activities
-            .FindAsync([request.Id], cancellationToken)
-            ?? throw new Exception("Activity not found");
+            .FindAsync([request.Id], cancellationToken);
+
+            if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
+
             context.Activities.Remove(activity);
-            await context.SaveChangesAsync(cancellationToken);
+            var result = await context.SaveChangesAsync(cancellationToken);
+            if (result == 0) return Result<Unit>.Failure("Failed to delete activity", 400);
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
